requirements.txt
discord.py>=2.3.2
anthropic>=0.7.0
openai>=1.3.0
google-generativeai>=0.3.0
redis>=5.0.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
